+++ tests/Base/SchemaWithItemsTest.php
@@ -4,19 +4,16 @@ declare(strict_types=1);
 
 namespace Yiisoft\Rbac\Cycle\Tests\Base;
 
-abstract class SchemaWithItems extends TestCase
+abstract class SchemaWithItemsTest extends TestCase
 {
     use SchemaTrait;
 
     protected static array $migrationsSubfolders = ['items'];
 
-    public function testCreateItemTables(): void
+    protected function checkTables(): void
     {
-        $this->runMigrations();;
-
         $this->checkItemsTable();
         $this->checkItemsChildrenTable();
-
         $this->assertFalse($this->getDatabase()->hasTable(self::$assignmentsTable));
     }
 }
