+++ tests/Mysql/SchemaWithItemsTest.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Rbac\Cycle\Tests\Mysql;
+
+class SchemaWithItemsTest extends \Yiisoft\Rbac\Cycle\Tests\Base\SchemaWithItemsTest
+{
+    use DatabaseTrait;
+}
