+++ tests/Pgsql/SchemaTest.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Rbac\Cycle\Tests\Pgsql;
+
+class SchemaTest extends \Yiisoft\Rbac\Cycle\Tests\Base\SchemaTest
+{
+    use DatabaseTrait;
+}
