+++ tests/Pgsql/SchemaWithAssignmentsTest.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Rbac\Cycle\Tests\Pgsql;
+
+class SchemaWithAssignmentsTest extends \Yiisoft\Rbac\Cycle\Tests\Base\SchemaWithAssignmentsTest
+{
+    use DatabaseTrait;
+}
