+++ tests/Sqlite/SchemaWithAssignmentsTest.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Rbac\Cycle\Tests\Sqlite;
+
+class SchemaWithAssignmentsTest extends \Yiisoft\Rbac\Cycle\Tests\Base\SchemaWithAssignmentsTest
+{
+    use DatabaseTrait;
+    use SchemaTrait;
+}
