+++ tests/Sqlite/SchemaWithItemsTest.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Rbac\Cycle\Tests\Sqlite;
+
+class SchemaWithItemsTest extends \Yiisoft\Rbac\Cycle\Tests\Base\SchemaWithItemsTest
+{
+    use DatabaseTrait;
+    use SchemaTrait;
+}
